fuking awosome

+++ src/containers/accounts/AccountsContainer.tsx
@@ -1,5 +1,8 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
+import { useNavigate } from "react-router-dom";
+import { ConfirmationDeleteDialog } from "../../components/dialogs/ConfirmationDeleteDialog";
+import { endroutes } from "../../constant/endroutes";
 import {
   deleteAccountAsync,
   fetchAccountsAsync,
@@ -17,19 +20,52 @@ const columns: any = [
 ];
 
 export const AccountsContainer = () => {
+  const navigate = useNavigate();
   const dispatch = useAppDispatch();
   const accounts = useSelector(selectAccounts);
 
-  const DeleteAccountAsync = (account: accountType) =>
-    dispatch(deleteAccountAsync(account));
+  // Redux Async Functions
+  const DeleteAccountAsync = (account: accountType, isDelete: boolean) => {
+    if (isDelete && account.id) {
+      dispatch(deleteAccountAsync(account));
+    } else
+      setShowConfirmationDialog({
+        show: true,
+        data: account,
+        title: account.name,
+        body: "are you sure you want to delete " + account.name,
+      });
+  };
+
+  const handleNavigateAccount = (key: string) =>
+    navigate(endroutes.account_statment(key).go);
 
   useEffect(() => {
     dispatch(fetchAccountsAsync());
   }, []);
 
-  const props = { accounts, columns, DeleteAccountAsync };
+  const [showConfirmationDialog, setShowConfirmationDialog] = useState<any>({
+    show: false,
+    data: {},
+    title: "",
+    body: "",
+  });
+
+  const props = {
+    accounts,
+    columns,
+    DeleteAccountAsync,
+    handleNavigateAccount,
+  };
   return (
     <>
+      <ConfirmationDeleteDialog
+        data={showConfirmationDialog}
+        setData={setShowConfirmationDialog}
+        handleSubmit={() =>
+          DeleteAccountAsync(showConfirmationDialog.data, true)
+        }
+      />
       <AccountDialog />
       <AccountsPage {...props} />
     </>
